move fuels2018 and fuels2030 to subfolders energy_demand and energy_production
