+++ generatorcore/fuels2018/energy_demand/energyDemand.py
@@ -1,0 +1,8 @@
+# pyright: strict
+
+from dataclasses import dataclass
+
+
+@dataclass(kw_only=True)
+class EnergyDemand:
+    energy: float
