+++ generatorcore/fuels2018/energy_production/fuelProduction.py
@@ -4,18 +4,6 @@ from dataclasses import dataclass
 
 
 @dataclass(kw_only=True)
-class F:
-    CO2e_combustion_based: float
-    CO2e_production_based: float
-    CO2e_total: float
-
-
-@dataclass(kw_only=True)
-class EnergyDemand:
-    energy: float
-
-
-@dataclass(kw_only=True)
 class FuelProduction:
     CO2e_production_based: float
     CO2e_production_based_per_MWh: float
