+++ generatorcore/fuels2018/f18.py
@@ -2,7 +2,9 @@
 
 from dataclasses import dataclass
 
-from .dataclasses import EnergyDemand, FuelProduction, TotalFuelProduction, F
+from .energy_demand import EnergyDemand
+from .energy_production import FuelProduction, TotalFuelProduction
+from .f import F
 
 
 @dataclass(kw_only=True)
@@ -13,12 +15,13 @@ class F18:
     d_i: EnergyDemand
     d_t: EnergyDemand
     d_a: EnergyDemand
+
     p_petrol: FuelProduction
     p_jetfuel: FuelProduction
     p_diesel: FuelProduction
     p_bioethanol: FuelProduction
     p_biodiesel: FuelProduction
     p_biogas: FuelProduction
-
     p: TotalFuelProduction
+
     f: F
